refactor: move model linter logic to separate module
