+++ web/src/ModelEditPage.js
@@ -18,12 +18,8 @@ import * as ModelBackend from "./backend/ModelBackend";
 import * as OrganizationBackend from "./backend/OrganizationBackend";
 import * as Setting from "./Setting";
 import i18next from "i18next";
-
 import {Controlled as CodeMirror} from "react-codemirror2";
-import "codemirror/lib/codemirror.css";
-import "codemirror/addon/lint/lint.css";
-import "codemirror/addon/lint/lint";
-import {newModel} from "casbin";
+import {createLinter} from "./utils/modelLinter";
 
 require("codemirror/mode/properties/properties");
 
@@ -92,44 +88,6 @@ class ModelEditPage extends React.Component {
       model: model,
     });
   }
-
-  checkModelSyntax = (modelText) => {
-    try {
-      const model = newModel(modelText);
-      if (!model.model.get("r") || !model.model.get("p") || !model.model.get("e")) {
-        throw new Error("Model is missing one or more required sections (r, p, or e)");
-      }
-      return null;
-    } catch (e) {
-      return e.message;
-    }
-  };
-
-  createLinter = (CodeMirror) => {
-    CodeMirror.registerHelper("lint", "properties", (text) => {
-      const error = this.checkModelSyntax(text);
-      if (error) {
-        const lineMatch = error.match(/line (\d+)/);
-        if (lineMatch) {
-          const lineNumber = parseInt(lineMatch[1], 10) - 1;
-          return [{
-            from: CodeMirror.Pos(lineNumber, 0),
-            to: CodeMirror.Pos(lineNumber, text.split("\n")[lineNumber].length),
-            message: error,
-            severity: "error",
-          }];
-        } else {
-          return [{
-            from: CodeMirror.Pos(0, 0),
-            to: CodeMirror.Pos(text.split("\n").length - 1),
-            message: error,
-            severity: "error",
-          }];
-        }
-      }
-      return [];
-    });
-  };
 
   renderModel() {
     return (
@@ -204,7 +162,7 @@ class ModelEditPage extends React.Component {
                   this.updateModelField("modelText", value);
                 }}
                 editorDidMount={(editor, value, cb) => {
-                  this.createLinter(editor.constructor);
+                  createLinter(editor.constructor);
                 }}
               />
             </div>
